Eliminate openscreen_unittests log spam, fix some gmock warnings.
Use ::testing::NiceMock and/or correct unit tests that were missing
EXPECT_CALL()'s that were actually relevant to the test. Also, raise the
default log level to kWarning (was kInfo); but all command-line tools
and demo apps explicitly set this back down to kInfo.

Background: Many of the tests were spamming the logs with "uninteresting
mock" warnings from GMock. Also, there was a ton of spam due to INFO
logging meant for the command-line tools. This made it difficult for
humans to find the logging output associated with test failures.
Change-Id: I2560a7324130727c25e0593b86d08b4753e0570e