Update patch set 1
Patch Set 1:

(4 comments)

Nice change, thanks!  A couple of substantial comments.

It's good to have nice, mockable interfaces in platform/api.  But we introduce the risk of losing ABI compatiblity when we introduce a vtable based interface between the platform and the embedder.  I suppose this is already a problem with the way we've structured other public APIs in the library.

We can fix this later by introducing C-style interfaces to proxy the C++ ones.  It's a little gross but might give us the best of both worlds (good testability and ABI stability).   I think this is what libassistant does - maybe WebRTC has a different approach.

I'll add a tracking item to figure out our ABI story to the sheet.  Not something to resolve  in this context.

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>