Update patch set 11
Patch Set 11:

(28 comments)

I had a whole bunch of "spit and polish" comments, and a few questions about some usage scenarios that don't seem to be covered (which may amount to "that's a feature request" or "it'll be the next CL"), and one usage scenario I don't think we'll ever have.

Very nice to have the documentation. It helped a lot when reviewing the code. :)

Patch-set: 11
Reviewer: Gerrit User 1002076 <1002076@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>