Revert double checkdeps.py call
Because of persisting confusion in having both PRESUBMIT.sh and
PRESUBMIT.py, a change recently added checkdeps.py to PRESUBMIT.sh, when
it's already in PRESUBMIT.py.  This change reverts that but also adds a
warning about running PRESUBMIT.sh.  This is just a stop-gap until
PRESUBMIT.sh is removed.

Bug: openscreen:43
Change-Id: I0e7b6f97b3afd416c1e20a3db66adbe9f6a10498