Refactor UDPSocket creation
This patch removes the Version parameter in the Create method for UDPSocket,
and moves the endpoint parameter from the Bind() method to the
constructor. This makes the socket creation clearer, since you need to
have knowledge of the endpoint you want to connect to anyway if you know
what version it is going to be.