Trace Logging: Support User Args (DO NOT REVIEW YET)
This is a larger CL showing why a smaller one is needed:
Smaller CL: https://chromium-review.googlesource.com/c/openscreen/+/1726710
I will update this one once the smaller one is merged in



Add support for user-defined tracing arguments

This change makes heavy use of templates, but there is no way around this
because we want to support user-defined arguments of any user-provided
type

In order to optimize performance, I use the same trick as in the previous
TraceLogging CL where a unique_ptr is allocated with stack memory and a
custom deleter. I've wrapped this in a class as part of this change to
hide some of the complexity.

User arguments are provided as follows:

  TRACE_SCOPED(category, name, "UserArg1", "value", "UserArg2", 1);

These new user arguments are optional, and any currently available logging is
still available.
Further information can be found in platform/api/trace_logging_unittest.cc

Due to the structure of this change, it would also be simple to add support for

  TRACE_SET_ARGUMENT("UserArg", value)

If support for both approaches is helpful. I've only implemented the one for
now.


I will update the documentation as a follow-up CL. I want to ensure there are
no large changes before making these changes

Bug: openscreen:38
Change-Id: Ia335a7f8ed23d17e04a3e86b2d459d792f89dcf9