Cast Streaming: RTP packetizing around the max packet limit.
Two tweaks, relating to the same issue:

1. Perform additional parsing checks for invalid packet IDs (i.e., when
the special 0xffff value is used in the wrong contexts).

2. Minor API change, when packetizing frames into RTP packets: Instead
of a DCHECK crashing the process, return -1 from

RtpPacketizer::ComputeNumberOfPackets() if a too-large frame cannot be
packetized. This is because the user of the Cast Streaming library isn't
expected to know the exact internal design limits of the protocol. We
will provide a relevant error code in upstream code.
Change-Id: I6cd96a442a574b53577e8218d6a0dcfbca376415