UdpSocket: Allow IPADDR_ANY and port 0 when Create()'ing.
Removes the restriction that all UdpSockets must be created with a known
local endpoint (addr+port), and provides a GetLocalEndpoint() method to
query what port is being auto-assigned by the operating system. This
allows for two socket use cases:

1. Allowing the operating system to auto-assign a free port, which is
useful for things like Cast Streaming senders, unit testing, etc.

2. Direct point-to-point Send(), where the client code does not care
what the local endpoint is (and is fine with the operating system auto-
assigning whatever it wants as the "return address").

Change-Id: I623d3294f1dfdf7c9cc11bbf90acd2c3503fbfec