Update patch set 4
Patch Set 4: Commit-Queue+1

(1 comment)

> Patch Set 4:
> 
> (1 comment)
> 
> Would a normal Observer pattern work for this use case?
> 
> UdpSocket::Observer {
>   OnCreated(const UdpSocket* socket);
>   OnDestroyed(const UdpSocket* socket);
> }

The reason I went with direct calls to NetworkRunner instead of an observer pattern is because:
1) We need a TaskRunner (which NetworkRunner implements) already passed to the UdpSocket, so that NetworkRunner can post callbacks when a read happens.
2) By allowing arbitrary external classes to implement UdpSocket::Observer, we add an extra level of indirection which, beyond complicating the code, will require us to maintain more complex code (ie: swapping out the observer, cases where the observer might be nullptr, etc). 3) Since we ALWAYS require the NetworkRunner to perform these actions (or risk a SegFault), I see no reason to allow the coder to accidentally NOT provide it

If it turned out later that we did need other observers beyond the NetworkRunner, I be all for updating to that pattern. But for now, I saw no advantage to doing so but multiple issues with it - so I tried to avoid it

Thoughts?

Patch-set: 4
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1