Udp Socket: Remove Deletion Callback
The approach here is to remove the socket-specific deletion callback and
replace it with an observer. The purpose of this change is to allow for
the eventual removal of the NetworkRunner class

Currently, we use NetworkRunner::ReadRepeatedly to signal to the
NetworkReader that it needs to start watching a given socket. After this
change, we can instead start watching a socket upon creation and stop
watching it when it's deleted. This will allow us to break the
dependency on the NetworkRunner, and remove the ReadRepeatedly /
CancelRead methods (making it just a normal TaskRunner), and then remove
the whole class.

The changes discussed in the paragraph above have a dependency on other
open CLs, though, so they cannot be included in this CL.

Change-Id: I84b6664d5ddf1dcc767e7888411922c0c7c5a5a1