Remove gcc bots
This change removes our test support for gcc for the following reasons:

 1. Building on gcc 8 doesn't work and doesn't appear to be trivial to
    fix.  This means local builds with gcc don't work, and upgrading our
    bots will break gcc support without fixes.
 2. We aren't required to support gcc to build in Chromium.
 3. In addition to missing flags and features due to the older version
    on the bots, there has also been a mysterious crash that also
    doesn't repro locally.

Given these reasons, gcc is costing more than it's helping at this
point.  This change doesn't remove the is_gcc flag from the GN files or
the bot recipe though, so it can continue to be improved in the
background if that ends up being desirable.

Bug: openscreen:40
Change-Id: I41142cb4370f469e0c8d185a49968b25f1d032ee