Update patch set 1
Patch Set 1: Code-Review+1

No objection to dropping GCC. However, it would be great to figure out the cause for the crash. We've had an issue that was showing up only on Mac build, and it turned out to be a real problem.

Patch-set: 1
Reviewer: Gerrit User 1336186 <1336186@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1