Update patch set 1
Patch Set 1:

> 2. I echo Brandon's point, that it has been tough to maintain support for gcc: The bots only provide really old compilers that *do* have glaring bugs (e.g., around noexcept and other relatively straightforward C++14 syntax/features). While newer gcc versions have resolved all these issues, if our infrastructure doesn't provide reasonably-new versions, we don't have the available manpower to work around those silly things.

We _do_ now have a path to an 18.04 bot with gcc ~8.2, but this led to other problems (noted in the bug and in my response to the next point).

> 3. There's mention that gcc-8 doesn't work: That's alarming. But also, it's only because of a crash in gmtime()/localtime()? Do we call those (outside of platform/api/time), or is this within platform/api/time code? Are we using the thread-unsafe versions of these functions in a multi-threaded context? It's surprising to see those functions crashing since that would break a ton of code for all gcc users: So, are we just removing the canary from the coal mine here? Maybe there's a legitimate bug in OpenScreen code we need to address here?

gcc-8 doesn't work because it complains that many defaulted constructors and operators are implicitly deleted because the exception specification is wrong.  You can look at the bug for the exact error I saw.

The gmtime/localtime crash is on 5.4.  We are using the non-reentrant versions in a single-threaded context (currently), though in the future we'd either use the reentrant ones or need to rewrite our own function (for precision reasons).

Patch-set: 1
Reviewer: Gerrit User 1215804 <1215804@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>