Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I won't +1 nor -1 this idea. My thoughts:
> 
> 1. I'm concerned that we're writing C++14-to-spec code that builds on Clang but not on gcc. Either that means Clang is permissive where it shouldn't be, or gcc has bugs that Clang doesn't (or vice versa), or MOST IMPORTANTLY that we are relying on undefined behavior. If it's for this last reason, then we should be fixing our code.

I don't think we'll know what needs fixing, until we have some way to use an up-to-date gcc in our CI.

In the meantime we'll have to rely on clang warnings to warn us of undefined behavior.
 
> 2. I echo Brandon's point, that it has been tough to maintain support for gcc: The bots only provide really old compilers that *do* have glaring bugs (e.g., around noexcept and other relatively straightforward C++14 syntax/features). While newer gcc versions have resolved all these issues, if our infrastructure doesn't provide reasonably-new versions, we don't have the available manpower to work around those silly things.

There are other working on/in Chromium who care about building with gcc:

https://crbug.com/993449
https://crbug.com/994581
https://crbug.com/994556

etc.

Maybe we can get a critical mass to get gcc more officially supported in the infrastructure.

> 3. There's mention that gcc-8 doesn't work: That's alarming. But also, it's only because of a crash in gmtime()/localtime()? Do we call those (outside of platform/api/time), or is this within platform/api/time code? Are we using the thread-unsafe versions of these functions in a multi-threaded context? It's surprising to see those functions crashing since that would break a ton of code for all gcc users: So, are we just removing the canary from the coal mine here? Maybe there's a legitimate bug in OpenScreen code we need to address here?

+1 

It would be helpful to understand what is going on with this, maybe file an issue with steps to reproduce so we can figure out if this is a compiler issue, code issue or libc issue.

But it shouldn't block turning down the gcc bots for us specifically.

Patch-set: 1
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>