Cast Streaming: Change EncodedFrame API to allow external buffers.
After discussion with one of the future customers of the Cast Streaming
Receiver, an API adjustment is needed: It's common for a platform to
need to provide a buffer that its own media decoder can access
efficiently. Thus, this change:

1. Modifies the EncodedFrame API to allow external buffers to be
provided for receiving encoded data.

2. Modifies FrameCrypto's helper struct (EncryptedFrame, a substruct of
EncodedFrame) to also own/manage its own buffer, and maintain the
EncodedFrame::data field to point to it.

Also, refreshed the comments around how to use the timestamp fields in
EncodedFrame since I'm already touching the file and it'ss in the same
theme of "adjusting the API our customers will use."

Bug: openscreen:55
Change-Id: I0858ce6d69addf822fd52dff38777bf76bc0bd95