UdpSocket: Enable/Disable Reading Flag
The goal of this CL is to prepare for a post-network-runner world
Currently, if a caller wants to disable reading on a socket, they call

NetworkRunner::CancelRead(...) and then re-initialize it with
NetworkRunner::ReadRepeatedly(...). Once these methods are removed and
callbacks are moved to UdpSocket::Client, we need a new mechanism to
prevent callbacks from firing unneccisarily. This flag provides 2
benefits:
  1) In StandAlone, this flag can be directly checked by the
     NetworkReader (in this CL, enable/disable reading is added to
     ReadRepeatedly and CancelRead - this will not be the case once
     those methods are removed in the follow-up UdpSocket Read update)
  2) Callers can call enable_reading and disable_reading as needed
     without needinf to track socket state locallym, which would add
     unneccisary overhead in many places throughout our code (see the
     UdpSocket: Update Reads CL)
Change-Id: I0fccb59d10e229b7b2584496f0124d5923e70045