Cast Streaming: SenderReportParser::Parse() exposes report ID.
The report ID is a value generated from the wire-format NTP timestamp
field in an RTCP packet. This is used by the Receiver in its Receiver
Reports as a way of marking it as a reply to a past specific Sender
Report, which is important for the Sender to be able to compute network
round-trip times.

Why we don't just shove this into the common RtcpSenderReport struct:
The report ID is *not* an input on the serialization side of things.
Instead, it is an artifact resulting from how the reference_time input
is converted into the RTCP wire format/field.

Change-Id: Ia1d63127ae5bb87aba052aa1b4f0a2a7beed714d