Cast Streaming: Receiver skeleton, demo app, and usage documentation.
Introduces the Cast Streaming Receiver's public API (with stubs), along
with:

1. Extensive class-level comments and usage documentation.

2. A demo app showing how a system is set up for receiving media streams
and decoding/playout.

3. A new Environment class, meant to be shared by future Sender and
Receiver components, which is a thin wrapper for integrating with the
OSP platform API. It will also be used to cleanly inject "fake"
dependencies throughout all Cast Streaming components during testing.

4. A new ReceiverPacketRouter class, which dispatches the received UDP
packets to an appropriate receiver instance and sets up the "path back
to the Sender" for RTCP packets.

Bug: openscreen:55
Change-Id: I131345cd5f07011f2955649ccb0c58c2e83d6f52