Connect Cast Streaming to UdpSocket.
Changes the cast_streaming::Environment to create/bind/own the UDP
socket for the session, rather than have it passed-in to its
constructor. This allows it to become a UdpSocket::Client to be able to
dispatch inbound UDP packets to the rest of the system. Also, added the
calls to Start/Stop packet receives (platform impl is a TODO).

Updated downstream interfaces to support a zero-copy path for the packet
data, now that the upstream API has settled on the platform::UdpPacket
class.

Bug: openscreen:56
Change-Id: Ic71518ab88fc39b45d526fb6396aaf740e16f921