Update patch set 5
Patch Set 5: Commit-Queue+1

(4 comments)

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > (4 comments)
> > 
> > High-level comments:
> > 
> > 1. I suggest eliminating the TaskRunnerThread class altogether, and put RunUntilStopped() right in the main() function where you want the app to run until it's time to quit. It's much simpler that using thread synch mechanisms to block, and makes it much clearer where main()'s blocking is supposed to happen (without the reader having to worry about any threading).
> > 
> 
> All of this was started by my suggestion to make it easier to use the TaskRunner correctly.  Previously, it looks fairly tricky, and it's code that would end up being copied N times.  
> 
> I'm not tied to the idea of having it block when going out of scope.  I'm fine with their being a RunUntilStopped() in the main() function.  I'm not tied to any particular implementation.  I just wanted something a little easier to use than what we had.
> 
> But if this ends up being too hard to pull off, we should just back up to where we were without this helping code and just have something a little big complex copied N times.  I don't want getting this perfect to block progress on other things.
> 

If we want to eliminate the TaskRunnerThread / NetworkRunnerThread classes and have RunUntilStopped() in code we can, but that is pretty much the state of the code prior to this change
I am not at all against that approach, just want to have everyone agree first

As far as blocking when it goes out of scope, our options seem to be:
1) Block on the thread with a Thread::join() call
2) Detatch the thread (as Mark suggested)
3) Do neither of the above

#1 has the issue of blocking at program exit time, and #3 will crash the program with an unhealthy exit code
I was under the impression #1 would be a good approach because this is its own object, so if users don't want to block they don't have to use it and can instead create a Task Runner manually, but if this isn't the approach we want how about #2?

> 
> 
> 
> 
> 
> > 2. The code being added to platform/impl does not implement any platform APIs/interfaces. Meaning, we would not expect our customers to have to implement these things for their app. So, they should probably go in the osp/demo directory alongside the demo app code.

The issue with putting this in the demo code is that it would need to be copied to a bunch of places - we need it in demo, embedder_demo, presumably in your new cast demo, etc...
(At least if I am remembering our deps rules correctly. Maybe we could add common/demo or similar?)

Any ideas?

Patch-set: 5
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+1