Cast Streaming Receiver: Add RTCP processing.
Adds the implementation that processes the inbound RTCP Sender Report
packets, as well as the periodic Receiver Reports back to the sender.

Unit testing revealed a pre-existing bug in the ReceiverPacketRouter,
where the Receiver SSRC was being used instead of the Sender SSRC to
dispatch packets.

Also, to avoid the possibility of the wrong Clock being used by
NTPTimeConverter (e.g., real versus FakeClock), its first constructor
argument is now required (not defaulted when unspecified).

Bug: openscreen:56
Change-Id: I7feeb04c6b6ee167768bb69d300f2c77a54727ec