Update patch set 3
Patch Set 3:

(11 comments)

So, if I understand this correctly, framer tries to read a message from a buffer that is written to elsewhere.

A span is passed to the constructor of the framer, but it only uses the pointer to the start of the data. Also, there's no bounds checking in the framer that indeed it can try to read from input buffer as much as Ingest claims there is. This seems very dangerous and unsafe to me.

Would it be possible to create a reusable max size buffer and pass it to the span, then update valid data size as the data is received? Or even better, create a reader on the buffer after the entire message was received (in a valid scenario).

Patch-set: 3
Reviewer: Gerrit User 1336186 <1336186@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>