Cast Streaming Receiver: Frame queuing and processing of RTP packets.
Adds the implementation that parses incoming RTP packets, and collects
them together in an internal "frame queue." Once a frame is complete, a
consumer is notified. Also adds the RequestKeyFrame() implementation,
which initiates the Picture Loss Indicator; and the code that resolves/
clears the PLI once a key frame is received.

In a later commit, the implementation that extracts from the frame queue
will be added.

Made a small change in rtp_time.h to avoid the use of std::round() in a
constexpr function, which was preventing the compiler from evaluating an
expression at compile time.

Bug: openscreen:56
Change-Id: I3f36dd6b0a95bf9d2626def7307f546d2d9a7eec