Unify Networking Thread Pt1: Class Definition
There is some complexity to both this design and why I have gone with
this approach. *I couldn't find a good place to explain it in the code
itself - please let me know if you have any suggestions*

The base idea is that we need to allow for StartProviding and
StopProviding calls to come while the Wait() loop is running. Because
the entire wait loop needs access to the providers_ array, if we lock
around that array, it's possible that the StartProviding and
StopProviding methods will never actually acquire the lock. If that
happens, they will block indefinitely.

To get around this, instead of modifying the array iself, I queue up
operations that will occur in the providers_changes_ queue. This is
only briefly locked by the Wait() call (as it checks for new changes),
so the StartProviding and StopProviding calls should not need to wait
long to acquire the lock for 2 reasons:
  - Due to empty_providers_block_, the Wait() loop will never never
    acquire the lock for a nontrivial amount of time
  - By definition of Provider::PerformNetworkingOperation, we expect
    the operation to take a "nontrivial amount of time"
Together, these should guarantee that there is enough time between the
Wait() loop releasing the lock and re-acquiring it in the following
iteration to allow StartProviding() and StopProviding() a change to
acquire it.

This approach, though, introduces the possibility that a removed
provider may be called during the Wait loop. To prevent this edge case
(and possible resulting segfault), the stop_providing_block_ condition
variable is used.

Once this approach has been looked over, I will examine what UTs can be
added. I suspect there may be some larger design changes, so I would
like to go through some Code Review first.

Change-Id: Ia1bea489f3983a591f8b67e74107559bff57838f