General SubscriberModel util class
Currently, we have 2 places where we need a subscriber model:
 - SocketHandleWaiter
 - UdpSocketReader
We are about to add 3 more:
 - NetworkReaderWriter (see separate CL)
 - TlsDataRouter x 2

Of these, SocketHandleWriter, NetworkReaderWriter, and TlsDataRouter
need the ability to wait on the deletion of its managed object type
until its 'Wait loop' is no longer running. This class provides a
single class to use in each of the above cases

More details:
The base idea is that we need to allow for Subscribe and Unsubscribe
calls to come while the subscribers() vector is being used. If we lock
around the vector, it's possible that the Subscribe and Unsubscribe
threads will starve. To get around this, instead of modifying the
vector iself, I queue up operations. This way, the mutex is only briefly
locked and the vector never needs to be copied.

This approach, though, introduces the possibility that a removed
provider may be called during the Wait loop. To prevent this edge case
(and possible resulting segfault), the stop_providing_block_ condition
variable is used. As pointed out by Yuri, "This design is 'inside-out'
from what it should be, in that Wait() operations should always be
interruptible when control operations need to alter the configuration of
the system."

In order to allow for cancellation, I've added
SubscriberModelCancellationToken. This cancellation token provides a
way for the caller to respond to an ongoing block and end ongoing
operations earlyas follows:
 - In SocketHandleWaiter, a fake socket can be used to exit the Select
   call early.
 - In TlsDataRouter, all further Read and Write calls can be postponed.
 - In NetworkReaderWriter, the cancellation can be propegated to all
   registed providers via a new operation in the Provider interface.
While this doesn't provide a system level guarantee that the call will
return quickly, it should allow for the blocked operation to continue
nearly instantly.

Change-Id: Idb39eceaa73216d69da0971cc7b3bf52de5c9f7f