Update patch set 1
Patch Set 1:

(12 comments)

On the one hand it is better to be explicit about dependencies and allow mocks to be injected by passing the context everywhere.

On the other hand it looks like the context is a natural singleton that could own other singleton-like objects like the task runner and the network loop.  So it might not be the end of the world to have a SetInstance()/GetInstance() API on it.   That would remove a lot of RuntimeContext parameters from throughout the code.

Let's flesh out the RuntimeContext API and implementation and see where that lands before deciding which way to go?

Patch-set: 1
CC: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>