RuntimeContext API + Posix Impl
Currently, all our code passes around a TaskRunner. The goal with this change
is to create a new object to be passed around in place of a TaskRunner.

The RuntimeContext provides a platform-specific hook to access creation and
destruction of UdpSocket, TlsConnection, and TlsConnectionFactory objects.
These allow us, in Posix, to begin monitoring these objects on the
Networking thread.

For more details:
 - High level info, see option 1 in design doc:
https://drive.google.com/open?id=1ggXKvL5kS5pAiXgXoOC10qwlCOReu8yrkdlYi_2O_No
 - Integration with existing code, see:
https://chromium-review.googlesource.com/c/openscreen/+/1832011

Change-Id: Ibd647981dba87b8e040ca26ec1a7ce3827fd9dde