Update patch set 3
Patch Set 3:

I think it's better to have classes accept task runner as a parameter to the constructors.

Thinking of an arbitrary class as a complete type, the class constructor should list all dependencies that are needed for the class to function and accept those as parameters, and create everything else that's needed in the constructor.

It's better if the constructor explicitly lists the direct dependencies instead of accepting an uber-object and getting dependencies from it (almost like service locator pattern/anti-pattern). One of the reasons why is this way the class states what it needs to function and does not care about how the caller provides the dependencies.

Using RuntimeContext as parameter adds unnecessary coupling between it and other classes, because RuntimeContext itself is not necessary for the class to function correctly.

Patch-set: 3
Reviewer: Gerrit User 1336186 <1336186@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>