Add frame agent config files
This patch is the beginning of some work I am planning for implementing
the libcast API with @miu. Ultimately, we will have AudioSender and
VideoSender classes, and are finishing up the prototype/design stage for
those classes.

The first step is getting the configuration files checked in.

These configs are largely from the cast_config.h file in media/cast:

https://osscs.corp.google.com/chromium/chromium/src/+/master:media/cast/cast_config.h

However, they have been refactored with better typing, naming, and
layout (hopefully).

Change-Id: I8c0df75fd50cf0adcbd92382960c2bc74edb0c82