Update patch set 10
Patch Set 10:

(2 comments)

> Patch Set 10:
> 
> (2 comments)
> 
> Just so I understand: 
> 
> 1. Each operation is responsible for enforcing the timeout internally; the operation loop is not going to interrupt a task that exceeds the timeout?

That is correct. It might not be safe for us to interrupt some operations (ie, if we interrupt TLS data read, that could lead to data loss. Or interrupting Write could put us in an unexpected state where the caller thinks it was written and it wasn't)
> 
> 2. An operation is guaranteed to be run no sooner than every min_loop_execution_time, but could be run much later depending on the timeout value, the number of operations, and whether any operation exceeds its timeout?

That is correct
The idea here is that different embedders may want this to behave differently. For instance, in the case of the NetworkLoop:
 - an embedder who plans to use a dedicated core for the network loop may want a min execution time of 0, so that the loop runs as a hot loop
 - an embedder running on a Raspberry Pi v1 would want a high min loop time so that they dont starve the other threads
For this specific case, we can surface the value as a caller parameter by making it an input parameter to PlatformClientPosix::Create(...) (see parallel CL)

Similarly, if timeout is enforced here instead of the caller, it's possible to arrive in states where data can never be read (ie, if SocketHandleReader has internal timeout k, and this class uses timeout k / 2, then we will kill off SocketHandleReader.Wait() before it can ever read any data)

Because the subscribers are given as part of the constructor (n of them) along with timeout T and max time M, the user could set M = T * n, but this approach leaves that up to the user

Patch-set: 10
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0