Add Mac OS X ScopedWakeLock implementation
This patch adds a Mac-only ScopedWakeLock implementation, using the IOPM
library. Note that a global is necessary to keep track of state. This is
done using a new LockState type that is privately held by
ScopedWakeLockMac.

This scoped wake lock will be used by Open Screen to ensure that receivers
do not fall asleep/screen dim while we are casting content.

Change-Id: Ia69dd754533d636f96dd97f45c2c9a15a7f146e5