PlatformClient: Use In Existing Classes
This change calls the PlatformClientPosix instance from posix classes
which need to register or deregister instance creation and destruction.

PlatformClient instances are provided to these classes in the ctor to
allow for dependency injection in UTs.
Access to the platform_client_ instance in each of these classes is
guarded by checking if (platform_client_) to ensure that the instance
exists. This is primarily for UnitTest purposes, as it would require
large changes to our UnitTests to remove these checks. This can be
accomplished in a follow-up CL, but it is outside the scope of this
initial CL

Change-Id: I0e8b764740428dd5ddbbc5f2115a5e81bc826aa4