DNS-SD: Constants
This class defines a few common structs to be used throughout DNS-SD.
The name is taken for consistency with cast/mdns/constants.h which
defines similar structs.

SrvKey and PtrKey are needed to act as the HashKey of unordered_map
instances for the querier implimentation in a follow up CL. For that
purpose, equality operators and hasn functions have been defined for
both.

DnsData is needed to store data as we recieve mDNS data over the wire.

If further context is required for this CL, LMK and I can upload a
single CL containing most/all of the querier code (for reference
purposes) to show how these will be used.

Note: In a future CL, the mDNS objects here should be moved to a
'common' folder which doesn't even need to be mDNS specific. For this
reason, this class refernces 'DNS' rather than mDNS specifically.
Change-Id: I736306aaa636841820bcc9eb14781accae89e972