DNS-SD: TxtRecord implementation
This is the implementation of the TxtRecord class as defined in CL:
https://chromium-review.googlesource.com/c/openscreen/+/1860922

txt_record is implemented under the hood using an std::unordered_map
with custom hash and equality operators to account for no casing. Key
validation is done as part of setting the key/flag.

the txt_record interface was changed because all methods must
immediately create an std::string (see iteration 1) in order to access
the unordered_map via key type std::string

Change-Id: I42152e2605d5af507c95226b33cab4ed184f6f93